fastapi/requirements.txt
fastapi
uvicorn
numpy>=1.24.4,<2.0
annotated-types==0.7.0
bcrypt==4.0.1
certifi==2025.10.5
charset-normalizer==3.4.4
ecdsa==0.19.1
greenlet==3.2.4
idna==3.11
passlib==1.7.4
psycopg2==2.9.11
pyasn1==0.6.1
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.44
SQLAlchemy-Utils==0.42.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
pydantic[email]
opencv-python-headless
redis==5.0.8
qrcode[pil]
cloudinary==1.41.0
ultralytics>=8.0.0
celery==5.4.0
flower==2.0.1
reportlab==4.0.7
python-magic==0.4.27
python-multipart